Exempt both halves of a declared pair from citation checking
Declaring `extends`/`extended_by` on a real corpus turned up a finding
the unit tests could not: `extended_by:` was read as a citation site, so
a completed edge naming a Proposed document reported "cited without
acknowledgement" — one finding per edge the fixer writes.

The forward field was already exempt, because a predecessor being retired
is what a line looks like. The converse states the same fact from the far
end, so exempting one and reporting the other is incoherent.

`relation_spans` was driven by chain configuration, which is the same
mistake the one-sided check had: which fields hold structure rather than
citations is a property of the relation, not of a view that walks it.
Moved to `relations.py` and widened to both halves of any declared pair.
