+++ record/devlog.d/2026/09/06/183842.md
@@ -107,3 +107,22 @@ Fired on the real corpus again, on the case that motivated it: retracted
 LIT-181's `extends: LIT-140` from the anthology's residual chain. Lint
 named it stale, `--fix` removed `LIT-181` from LIT-140's `extended_by:`
 and left `LIT-151` in place, lint went clean, second run a no-op.
+
+## And the bug that found
+
+Declaring `extends`/`extended_by` on the real corpus turned one lint
+finding red that had been green: `extended_by:` was being read as a
+*citation site*, so a completed edge pointing at a `Proposed` document
+reported "cited without acknowledgement". The forward field was already
+exempt — a predecessor being retired is what a line looks like — and the
+converse says the same fact from the far end, so exempting one and
+reporting the other hands back a finding for every edge the fixer writes.
+
+`relation_spans` was chain-driven, which is the same incoherence the
+check itself had: it belongs to the relation, not to the view. Moved to
+`relations.py` and widened to both halves of any declared pair.
+
+Worth noting how it surfaced. The unit tests were green; what caught it
+was running the change against 387 real documents and reading the diff of
+the generated report, not the code. That is the second finding on this
+feature that only the real corpus produced.
